+++ compiler/native/squel/test.cpp
@@ -47,12 +47,6 @@
 
 int main(int argc, char* argv[])
 {
-	var __NJS_ARGS = __NJS_Create_Array();
-	for( int i = 0; i < argc; i++)
-	{
-		__NJS_Object_Set(i, __NJS_VAR(argv[i]), __NJS_ARGS);
-	}
-
 	{INIT}
 
 	{CODE}