+++ example/arduino-led.js
@@ -4,7 +4,7 @@ var PIN_LED = arduino.led_builtin;
 
 function setup()
 {
-	 arduino.pinMode(PIN_LED, arduino.output);
+	arduino.pinMode(PIN_LED, arduino.output);
 }
 
 function loop()
@@ -12,5 +12,5 @@ function loop()
 	arduino.digitalWrite(PIN_LED, arduino.high);
 	arduino.delay(1000);
 	arduino.digitalWrite(PIN_LED, arduino.low);
-	arduino.delay(1000);
+	arduino.delay(500);
 }