+++ example/math.js
@@ -3,3 +3,17 @@ console.log("Math.pow(10, 2): " + Math.pow(10, 2));
 console.log("Math.pow(10.00, -2): " + Math.pow(10.00, -2));
 console.log("PI: " + Math.PI + " - " + "type: " + typeof Math.PI);
 console.log("PI * 10: " + Math.PI * 10 + " - " + "type: " + typeof Math.PI);
+
+console.log("-------------");
+
+console.log("var i = 1;");
+var i = 1;
+
+console.log("i++");
+i++;
+
+console.log("i should be 2: " + i);
+
+console.log("i *= 2");
+i *= 2;
+console.log("i should be 4: " + i);